Just created data/avro and moved AvroDataReader into it + spark 1.6.3
