+++ photon-client/src/main/scala/com/linkedin/photon/ml/cli/game/scoring/Driver.scala
@@ -26,7 +26,8 @@ import com.linkedin.photon.ml.avro.data.ScoreProcessingUtils
 import com.linkedin.photon.ml.avro.model.ModelProcessingUtils
 import com.linkedin.photon.ml.cli.game.GAMEDriver
 import com.linkedin.photon.ml.constants.StorageLevel
-import com.linkedin.photon.ml.data.{AvroDataReader, GameConverters, GameDatum, KeyValueScore}
+import com.linkedin.photon.ml.data.avro.AvroDataReader
+import com.linkedin.photon.ml.data.{GameConverters, GameDatum, KeyValueScore}
 import com.linkedin.photon.ml.evaluation.{EvaluatorFactory, EvaluatorType}
 import com.linkedin.photon.ml.util._
